backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
aiosqlite==0.20.0
websockets==12.0
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.7.1
aiohttp==3.9.5
